+++ package.json
@@ -27,11 +27,11 @@
     "build": "nuxt build",
     "build-storybook": "build-storybook -s ./.storybook/assets",
     "build:test": "cross-env NODE_ENV=test nuxt build",
+    "cache": "node src/cachers/index.js",
     "dev": "nuxt",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
     "lint:fix": "npm run lint -- --fix",
-    "serverless": "node src/serverless/index.js",
     "start": "nuxt start",
     "start:inspect": "node --inspect node_modules/.bin/nuxt start",
     "storybook": "start-storybook -p 6006 -s ./.storybook/assets",
