+++ src/server-middleware/api/version.js
@@ -1,2 +1,2 @@
-console.log('require.main.filename', require.main.filename);
+// Outputs as plain text the version of the app that is running
 module.exports = (req, res) => res.send(require('../../../package').version);
